Combined methods for setting the color of the height map, moved class into gpuHeightMap to slowly start deprecating the CPU one
