+++ ihmc-footstep-planning/src/main/java/us/ihmc/footstepPlanning/bodyPath/BodyPathCollisionDetector.java
@@ -3,7 +3,7 @@ package us.ihmc.footstepPlanning.bodyPath;
 import gnu.trove.list.array.TIntArrayList;
 import us.ihmc.euclid.tools.EuclidCoreTools;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 
 /* package-private */
 class BodyPathCollisionDetector
