+++ ihmc-footstep-planning/src/main/java/us/ihmc/footstepPlanning/bodyPath/HeightMapLeastSquaresNormalCalculator.java
@@ -9,7 +9,7 @@ import us.ihmc.euclid.tuple3D.interfaces.UnitVector3DBasics;
 import us.ihmc.footstepPlanning.graphSearch.FootstepPlannerEnvironmentHandler;
 import us.ihmc.footstepPlanning.polygonSnapping.HeightMapPolygonSnapper;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 
 import java.util.Arrays;
