+++ ihmc-footstep-planning/src/main/java/us/ihmc/footstepPlanning/swing/HeightMapCollisionDetector.java
@@ -1,7 +1,6 @@
 package us.ihmc.footstepPlanning.swing;
 
 import org.ejml.data.DMatrixRMaj;
-import us.ihmc.euclid.referenceFrame.interfaces.FrameBox3DReadOnly;
 import us.ihmc.euclid.shape.collision.EuclidShape3DCollisionResult;
 import us.ihmc.euclid.shape.primitives.interfaces.Box3DReadOnly;
 import us.ihmc.euclid.tools.EuclidCoreTools;
@@ -11,7 +10,7 @@ import us.ihmc.euclid.tuple3D.interfaces.Point3DBasics;
 import us.ihmc.euclid.tuple3D.interfaces.Point3DReadOnly;
 import us.ihmc.euclid.tuple3D.interfaces.Vector3DReadOnly;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 
 public class HeightMapCollisionDetector
 {
