+++ ihmc-footstep-planning/src/test/java/us/ihmc/footstepPlanning/tools/PlanarRegionToHeightMapConverterTest.java
@@ -14,7 +14,7 @@ import us.ihmc.robotics.geometry.PlanarRegion;
 import us.ihmc.robotics.geometry.PlanarRegionsList;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
 import us.ihmc.sensorProcessing.heightMap.HeightMapMessageTools;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 
 import java.util.ArrayList;
 import java.util.List;
