+++ ihmc-footstep-planning/src/visualizers/java/us/ihmc/footstepPlanning/ui/viewers/HeightMapVisualizer.java
@@ -19,7 +19,7 @@ import us.ihmc.idl.IDLSequence;
 import us.ihmc.javaFXToolkit.shapes.JavaFXMultiColorMeshBuilder;
 import us.ihmc.javaFXToolkit.shapes.TextureColorAdaptivePalette;
 import us.ihmc.robotics.geometry.PlanarRegionsList;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
