+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/perception/RDXRemoteHeightMapPanel.java
@@ -6,12 +6,12 @@ import org.bytedeco.javacpp.BytePointer;
 import perception_msgs.msg.dds.HeightMapMessage;
 import perception_msgs.msg.dds.HeightMapStateRequestMessage;
 import us.ihmc.communication.PerceptionAPI;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 import us.ihmc.perception.heightMap.HeightMapAPI;
 import us.ihmc.communication.ros2.ROS2Helper;
 import us.ihmc.rdx.imgui.RDXPanel;
 import us.ihmc.rdx.imgui.ImGuiUniqueLabelMap;
 import us.ihmc.rdx.ui.ImGuiRemoteROS2StoredPropertySetGroup;
-import us.ihmc.rdx.ui.graphics.RDXGridMapGraphic;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
 import us.ihmc.sensorProcessing.heightMap.HeightMapFilterParameters;
 import us.ihmc.sensorProcessing.heightMap.HeightMapMessageTools;
@@ -77,7 +77,7 @@ public class RDXRemoteHeightMapPanel
          {
             int row = size - x - 1;
             int col = size - y - 1;
-            Color color = RDXGridMapGraphic.computeColorFromHeight(heightMapData.getHeightAt(x, y));
+            Color color = HeightMapTools.computeGDXColorFromHeight(heightMapData.getHeightAt(x, y));
             BytePointer pixel = heightMapPanel.getImage().ptr(row, col);
             int r = (int) (color.r * 255);
             int g = (int) (color.g * 255);
