+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/ui/graphics/RDXGridMapGraphic.java
@@ -13,14 +13,13 @@ import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.Pool;
 import org.lwjgl.opengl.GL41;
 import perception_msgs.msg.dds.HeightMapMessage;
-import us.ihmc.commons.InterpolationTools;
 import us.ihmc.euclid.transform.RigidBodyTransform;
 import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.euclid.tuple3D.interfaces.Point3DReadOnly;
 import us.ihmc.euclid.tuple3D.interfaces.Vector3DReadOnly;
 import us.ihmc.rdx.mesh.RDXIDMappedColorFunction;
 import us.ihmc.rdx.mesh.RDXMultiColorMeshBuilder;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 import us.ihmc.tools.thread.MissingThreadTools;
 import us.ihmc.tools.thread.ResettableExceptionHandlingExecutorService;
 
@@ -233,7 +232,7 @@ public class RDXGridMapGraphic implements RenderableProvider
 
                Color color;
                if (computeColorFromHeight)
-                  color = computeColorFromHeight(0.5 * (maxHeight + minHeight));
+                  color = HeightMapTools.computeGDXColorFromHeight(0.5 * (maxHeight + minHeight));
                else
                   color = idColorFunction.getColor((int) id);
 
@@ -283,7 +282,7 @@ public class RDXGridMapGraphic implements RenderableProvider
 
             Color color;
             if (computeColorFromHeight)
-               color = computeColorFromHeight(0.5 * (maxHeight + minHeight));
+               color = HeightMapTools.computeGDXColorFromHeight(0.5 * (maxHeight + minHeight));
             else
                color = idColorFunction.getColor((int) id);
 
@@ -307,7 +306,11 @@ public class RDXGridMapGraphic implements RenderableProvider
    {
       RDXMultiColorMeshBuilder groundMeshBuilder = new RDXMultiColorMeshBuilder();
       double renderedGroundPlaneHeight = 0.005;
-      groundMeshBuilder.addBox(gridSizeXy, gridSizeXy, renderedGroundPlaneHeight, new Point3D(gridCenterX, gridCenterY, groundHeight), computeColorFromHeight(groundHeight));
+      groundMeshBuilder.addBox(gridSizeXy,
+                               gridSizeXy,
+                               renderedGroundPlaneHeight,
+                               new Point3D(gridCenterX, gridCenterY, groundHeight),
+                               HeightMapTools.computeGDXColorFromHeight(groundHeight));
 
       return groundMeshBuilder;
    }
@@ -364,64 +367,6 @@ public class RDXGridMapGraphic implements RenderableProvider
       }
 
       return height;
-   }
-
-   public static Color computeColorFromHeight(double height)
-   {
-      // Using interpolation between key color points
-      double r = 0, g = 0, b = 0;
-      double redR = 1.0, redG = 0.0, redB = 0.0;
-      double magentaR = 1.0, magentaG = 0.0, magentaB = 1.0;
-      double orangeR = 1.0, orangeG = 200.0 / 255.0, orangeB = 0.0;
-      double yellowR = 1.0, yellowG = 1.0, yellowB = 0.0;
-      double blueR = 0.0, blueG = 0.0, blueB = 1.0;
-      double greenR = 0.0, greenG = 1.0, greenB = 0.0;
-      double gradientSize = 0.2;
-      double gradientLength = 1.0;
-      double alpha = height % gradientLength;
-      if (alpha < 0)
-         alpha = 1 + alpha;
-      while (alpha > 5 * gradientSize)
-         alpha -=  5 * gradientSize;
-
-      if (alpha <= gradientSize * 1)
-      {
-         r = InterpolationTools.linearInterpolate(magentaR, blueR, (alpha) / gradientSize);
-         g = InterpolationTools.linearInterpolate(magentaG, blueG, (alpha) / gradientSize);
-         b = InterpolationTools.linearInterpolate(magentaB, blueB, (alpha) / gradientSize);
-      }
-      else if (alpha <= gradientSize * 2)
-      {
-         r = InterpolationTools.linearInterpolate(blueR, greenR, (alpha - gradientSize * 1) / gradientSize);
-         g = InterpolationTools.linearInterpolate(blueG, greenG, (alpha - gradientSize * 1) / gradientSize);
-         b = InterpolationTools.linearInterpolate(blueB, greenB, (alpha - gradientSize * 1) / gradientSize);
-      }
-      else if (alpha <= gradientSize * 3)
-      {
-         r = InterpolationTools.linearInterpolate(greenR, yellowR, (alpha - gradientSize * 2) / gradientSize);
-         g = InterpolationTools.linearInterpolate(greenG, yellowG, (alpha - gradientSize * 2) / gradientSize);
-         b = InterpolationTools.linearInterpolate(greenB, yellowB, (alpha - gradientSize * 2) / gradientSize);
-      }
-      else if (alpha <= gradientSize * 4)
-      {
-         r = InterpolationTools.linearInterpolate(yellowR, orangeR, (alpha - gradientSize * 3) / gradientSize);
-         g = InterpolationTools.linearInterpolate(yellowG, orangeG, (alpha - gradientSize * 3) / gradientSize);
-         b = InterpolationTools.linearInterpolate(yellowB, orangeB, (alpha - gradientSize * 3) / gradientSize);
-      }
-      else if (alpha <= gradientSize * 5)
-      {
-         r = InterpolationTools.linearInterpolate(orangeR, redR, (alpha - gradientSize * 4) / gradientSize);
-         g = InterpolationTools.linearInterpolate(orangeG, redG, (alpha - gradientSize * 4) / gradientSize);
-         b = InterpolationTools.linearInterpolate(orangeB, redB, (alpha - gradientSize * 4) / gradientSize);
-      }
-      else
-      {
-         throw new RuntimeException("no valid color");
-      }
-
-      if (r == 0.0 && g == 0.0 && b == 0.0)
-         throw new RuntimeException("Shouldn't return black.)");
-      return new Color((float) r, (float) g, (float) b, 1.0f);
    }
 
    @Override
