+++ ihmc-perception/src/main/java/us/ihmc/perception/gpuHeightMap/HeightMapTools.java
@@ -1,0 +1,195 @@
+package us.ihmc.perception.gpuHeightMap;
+
+import us.ihmc.commons.InterpolationTools;
+import com.badlogic.gdx.graphics.Color;
+
+/**
+ * Height map indexing tools. The height map spans a square region and is parametrized by the following values:
+ * - A discretization value
+ * - The grid size, i.e. side length of the square region it covers
+ * - Grid center, an xy coordinate which is the middle of the grid
+ *
+ * Cells are indexed two ways:
+ * - A unique integer key, which is zero-indexed and starts at the corner of the grid which is the negative-most x and y coordinates.
+ * - An (x,y) integer index pair, which is zero at the negative-most cell along each axis
+ */
+public class HeightMapTools
+{
+   /**
+    * The xy-indices of the center of the grid.
+    */
+   public static int computeCenterIndex(double gridSize, double resolution)
+   {
+      return (int) Math.round(0.5 * gridSize / resolution);
+   }
+
+   public static int coordinateToKey(double x, double y, double xCenter, double yCenter, double resolution, int centerIndex)
+   {
+      int xIndex = coordinateToIndex(x, xCenter, resolution, centerIndex);
+      int yIndex = coordinateToIndex(y, yCenter, resolution, centerIndex);
+      return indicesToKey(xIndex, yIndex, centerIndex);
+   }
+
+   public static double keyToXCoordinate(int key, double xCenter, double resolution, int centerIndex)
+   {
+      int xIndex = keyToXIndex(key, centerIndex);
+      return indexToCoordinate(xIndex, xCenter, resolution, centerIndex);
+   }
+
+   public static double keyToYCoordinate(int key, double yCenter, double resolution, int centerIndex)
+   {
+      int yIndex = keyToYIndex(key, centerIndex);
+      return indexToCoordinate(yIndex, yCenter, resolution, centerIndex);
+   }
+
+   public static int coordinateToIndex(double coordinate, double gridCenter, double resolution, int centerIndex)
+   {
+      return (int) Math.round((coordinate - gridCenter) / resolution) + centerIndex;
+   }
+
+   public static double indexToCoordinate(int index, double gridCenter, double resolution, int centerIndex)
+   {
+      return (index - centerIndex) * resolution + gridCenter;
+   }
+
+   public static int keyToXIndex(int key, int centerIndex)
+   {
+      return key % (2 * centerIndex + 1);
+   }
+
+   public static int keyToYIndex(int key, int centerIndex)
+   {
+      return key / (2 * centerIndex + 1);
+   }
+
+   public static int indicesToKey(int xIndex, int yIndex, int centerIndex)
+   {
+      return xIndex + yIndex * (2 * centerIndex + 1);
+   }
+
+   public static int getIndexFromCoordinates(double coordinate, float resolution, int offset)
+   {
+      return (int) (coordinate * resolution + offset);
+   }
+
+   public static double getCoordinateFromIndex(int index, double resolution, int offset)
+   {
+      return (index - offset) / resolution;
+   }
+
+   public static Color computeGDXColorFromHeight(double height)
+   {
+      // Using interpolation between key color points
+      double r = 0, g = 0, b = 0;
+      double redR = 1.0, redG = 0.0, redB = 0.0;
+      double magentaR = 1.0, magentaG = 0.0, magentaB = 1.0;
+      double orangeR = 1.0, orangeG = 200.0 / 255.0, orangeB = 0.0;
+      double yellowR = 1.0, yellowG = 1.0, yellowB = 0.0;
+      double blueR = 0.0, blueG = 0.0, blueB = 1.0;
+      double greenR = 0.0, greenG = 1.0, greenB = 0.0;
+      double gradientSize = 0.2;
+      double gradientLength = 1.0;
+      double alpha = height % gradientLength;
+      if (alpha < 0)
+         alpha = 1 + alpha;
+      while (alpha > 5 * gradientSize)
+         alpha -=  5 * gradientSize;
+
+      if (alpha <= gradientSize * 1)
+      {
+         r = InterpolationTools.linearInterpolate(magentaR, blueR, (alpha) / gradientSize);
+         g = InterpolationTools.linearInterpolate(magentaG, blueG, (alpha) / gradientSize);
+         b = InterpolationTools.linearInterpolate(magentaB, blueB, (alpha) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 2)
+      {
+         r = InterpolationTools.linearInterpolate(blueR, greenR, (alpha - gradientSize * 1) / gradientSize);
+         g = InterpolationTools.linearInterpolate(blueG, greenG, (alpha - gradientSize * 1) / gradientSize);
+         b = InterpolationTools.linearInterpolate(blueB, greenB, (alpha - gradientSize * 1) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 3)
+      {
+         r = InterpolationTools.linearInterpolate(greenR, yellowR, (alpha - gradientSize * 2) / gradientSize);
+         g = InterpolationTools.linearInterpolate(greenG, yellowG, (alpha - gradientSize * 2) / gradientSize);
+         b = InterpolationTools.linearInterpolate(greenB, yellowB, (alpha - gradientSize * 2) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 4)
+      {
+         r = InterpolationTools.linearInterpolate(yellowR, orangeR, (alpha - gradientSize * 3) / gradientSize);
+         g = InterpolationTools.linearInterpolate(yellowG, orangeG, (alpha - gradientSize * 3) / gradientSize);
+         b = InterpolationTools.linearInterpolate(yellowB, orangeB, (alpha - gradientSize * 3) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 5)
+      {
+         r = InterpolationTools.linearInterpolate(orangeR, redR, (alpha - gradientSize * 4) / gradientSize);
+         g = InterpolationTools.linearInterpolate(orangeG, redG, (alpha - gradientSize * 4) / gradientSize);
+         b = InterpolationTools.linearInterpolate(orangeB, redB, (alpha - gradientSize * 4) / gradientSize);
+      }
+      else
+      {
+         throw new RuntimeException("no valid color");
+      }
+
+      if (r == 0.0 && g == 0.0 && b == 0.0)
+         throw new RuntimeException("Shouldn't return black.)");
+      return new Color((float) r, (float) g, (float) b, 1.0f);
+   }
+
+   public static javafx.scene.paint.Color computeJavaFXColorFromHeight(double height)
+   {
+      // Using interpolation between key color points
+      double r = 0, g = 0, b = 0;
+      double redR = 1.0, redG = 0.0, redB = 0.0;
+      double magentaR = 1.0, magentaG = 0.0, magentaB = 1.0;
+      double orangeR = 1.0, orangeG = 200.0 / 255.0, orangeB = 0.0;
+      double yellowR = 1.0, yellowG = 1.0, yellowB = 0.0;
+      double blueR = 0.0, blueG = 0.0, blueB = 1.0;
+      double greenR = 0.0, greenG = 1.0, greenB = 0.0;
+      double gradientSize = 0.2;
+      double gradientLength = 1.0;
+      double alpha = height % gradientLength;
+      if (alpha < 0)
+         alpha = 1 + alpha;
+      while (alpha > 5 * gradientSize)
+         alpha -= 5 * gradientSize;
+
+      if (alpha <= gradientSize * 1)
+      {
+         r = InterpolationTools.linearInterpolate(magentaR, blueR, (alpha) / gradientSize);
+         g = InterpolationTools.linearInterpolate(magentaG, blueG, (alpha) / gradientSize);
+         b = InterpolationTools.linearInterpolate(magentaB, blueB, (alpha) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 2)
+      {
+         r = InterpolationTools.linearInterpolate(blueR, greenR, (alpha - gradientSize * 1) / gradientSize);
+         g = InterpolationTools.linearInterpolate(blueG, greenG, (alpha - gradientSize * 1) / gradientSize);
+         b = InterpolationTools.linearInterpolate(blueB, greenB, (alpha - gradientSize * 1) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 3)
+      {
+         r = InterpolationTools.linearInterpolate(greenR, yellowR, (alpha - gradientSize * 2) / gradientSize);
+         g = InterpolationTools.linearInterpolate(greenG, yellowG, (alpha - gradientSize * 2) / gradientSize);
+         b = InterpolationTools.linearInterpolate(greenB, yellowB, (alpha - gradientSize * 2) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 4)
+      {
+         r = InterpolationTools.linearInterpolate(yellowR, orangeR, (alpha - gradientSize * 3) / gradientSize);
+         g = InterpolationTools.linearInterpolate(yellowG, orangeG, (alpha - gradientSize * 3) / gradientSize);
+         b = InterpolationTools.linearInterpolate(yellowB, orangeB, (alpha - gradientSize * 3) / gradientSize);
+      }
+      else if (alpha <= gradientSize * 5)
+      {
+         r = InterpolationTools.linearInterpolate(orangeR, redR, (alpha - gradientSize * 4) / gradientSize);
+         g = InterpolationTools.linearInterpolate(orangeG, redG, (alpha - gradientSize * 4) / gradientSize);
+         b = InterpolationTools.linearInterpolate(orangeB, redB, (alpha - gradientSize * 4) / gradientSize);
+      }
+      else
+      {
+         throw new RuntimeException("no valid color");
+      }
+
+      if (r == 0.0 && g == 0.0 && b == 0.0)
+         throw new RuntimeException("Shouldn't return black.)");
+      return new javafx.scene.paint.Color((float) r, (float) g, (float) b, 1.0f);
+   }
+}
