+++ ihmc-perception/src/main/java/us/ihmc/perception/gpuHeightMap/SimpleGPUHeightMap.java
@@ -12,7 +12,6 @@ import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.euclid.tuple3D.Vector3D;
 import us.ihmc.euclid.tuple3D.interfaces.Point3DBasics;
 import us.ihmc.euclid.tuple3D.interfaces.Vector3DBasics;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
 
 public class SimpleGPUHeightMap
 {
