+++ ihmc-perception/src/main/java/us/ihmc/perception/heightMap/HeightMapUpdater.java
@@ -18,6 +18,7 @@ import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.euclid.tuple3D.interfaces.Point3DReadOnly;
 import us.ihmc.idl.serializers.extra.JSONSerializer;
 import us.ihmc.log.LogTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 import us.ihmc.sensorProcessing.heightMap.*;
 import us.ihmc.tools.property.StoredPropertySet;
