+++ ihmc-perception/src/main/java/us/ihmc/perception/heightMap/TerrainMapData.java
@@ -11,7 +11,7 @@ import us.ihmc.perception.gpuHeightMap.RapidHeightMapExtractor;
 import us.ihmc.perception.steppableRegions.SnapResult;
 import us.ihmc.robotics.geometry.LeastSquaresZPlaneFitter;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
-import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
+import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 
 import java.util.ArrayList;
