+++ ihmc-sensor-processing/build.gradle.kts
@@ -29,6 +29,7 @@ mainDependencies {
 
    api("us.ihmc:euclid-frame:0.21.0")
    api("us.ihmc:ihmc-communication:source")
+   api("us.ihmc:ihmc-perception:source")
    api("us.ihmc:ihmc-java-toolkit:source")
    api("us.ihmc:ihmc-robotics-toolkit:source")
    api("us.ihmc:ihmc-robot-models:source")
